Do not collapse trivial connections in Paths
fixes #131.

i.e., do not reglue bits in a ContourComponent with collapsed edges. If
we do so we loose vertices in the contour. Also this leads to trouble in
the IET induction process: When a SaddleConnection is found later, it
might cross such a vertex, therefore triggering an assertion since
saddle connections are not allowed to cross a marked vertex (even if
its total angle is 2π.)