Fix serialization of Vertex
I am not convinced that this is the right way to do this here. On the one
hand, you don't want to promise a certain implementation detail, on the
other hand, you don't want to make something available in the
serialization that is not available otherwise. Maybe we should go for a
pattern where things serialize to Any instead which contains a few
whitelisted types.