fixed Chain::operator==
this now really feels like a bit of a hack since we are using the
assumption that all the other operations on the coefficientss, such as
addition, do not care about denormalized inputs. It seems logical that
they don't but we probably should not make that assumption on FLINT's
implementaion?