Remove smart pointers from API
having the separation between stack and heap objects is weird and seems
a bit random in the API. Basically, it exposes underlying implementation
details, e.g., a SaddleConnection needs to hold on to the surface it's
defined on and therefore holds a smart pointer to it. Consequently, all
surfaces are passed by shared_ptr. A downside of this is that we
essentially need to discourage people from calling the actual
constructors of these objects (or hide them away completely) since
everything needs to be created through make_shared. Futhermore, cppyy
has (understandably) trouble with our methods that return unique_ptr
which they merely do to emphasize that they return a fresh unrelated
copy of a surface.

So, let's hide these memory issue away in the PIMPL so that surfaces
feel like any other stack allocated object (even though they are now
always shared pointers behind the scenes.)

Also renamed some names that had been bothering for a while such as slot
which is now called slit (fixes #114) and fromEdge which is now called
fromHalfEdge.