Streamline use of noexcept
Since we really would like to see proper exceptions, in particular when
calling into C++ code from Python, we removed noexcept in most places.
It had originally been added due to a misunderstanding about performance
implications. We now left it in the places where it actually makes a
difference for performance, see
https://akrzemi1.wordpress.com/2011/06/10/using-noexcept/.