Expose CompareSlopes() in a more convenient way in Python
We opted for .slopes() to return saddle connections instead of vectors
for no strong reason. The scaling of the vectors might otherwise
confusing. Also, we do not win anything performance-wise by hiding that
information from the user…and the user just needs to call .vector() to
get the vector representation of the slope.